Added thrift installed check in build.sh

+++ build.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
+command -v thrift >/dev/null 2>&1 || { echo "thrift is not installed" >&2; exit 1; }
+
 rm -rf src
 mkdir src
 thrift -gen php:oop -out src/ vendor/elasticsearch/elasticsearch-transport-thrift/elasticsearch.thrift
-
+echo "Elasticsearch classes were regenerated in src folder"
